+++ STutor/src/main/java/ch/unibe/ese/controller/pojos/LectureForm.java
@@ -1,5 +1,6 @@
 package ch.unibe.ese.controller.pojos;
 
+
 /**
  * 
  * @author Christian Zürcher
@@ -11,30 +12,37 @@ public class LectureForm {
 	private String name;
 	private Long subject;
 	private Long university;
-	private float grade;
-	
+	private double grade;
+
 	public String getName() {
 		return name;
 	}
+
 	public void setName(String name) {
 		this.name = name;
 	}
+
 	public Long getSubject() {
 		return subject;
 	}
+
 	public void setSubject(Long subject) {
 		this.subject = subject;
 	}
+
 	public Long getUniversity() {
 		return university;
 	}
+
 	public void setUniversity(Long university) {
 		this.university = university;
 	}
-	public float getGrade() {
+
+	public double getGrade() {
 		return grade;
 	}
-	public void setGrade(float grade) {
+
+	public void setGrade(double grade) {
 		this.grade = grade;
 	}
